+++ internal/cloud/gcp/utils.go
@@ -1,0 +1,18 @@
+package gcp
+
+import (
+	"github.com/hashicorp/errwrap"
+	"google.golang.org/api/googleapi"
+)
+
+func isConflictError(err error) bool {
+	if e, ok := err.(*googleapi.Error); ok && (e.Code == 409 || e.Code == 412) {
+		return true
+	} else if !ok && errwrap.ContainsType(err, &googleapi.Error{}) {
+		e := errwrap.GetType(err, &googleapi.Error{}).(*googleapi.Error)
+		if e.Code == 409 || e.Code == 412 {
+			return true
+		}
+	}
+	return false
+}
